backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
firebase-admin==6.3.0
langdetect==1.0.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator
pydantic-settings
deep-translator==1.11.4
textblob==0.17.1